backend/requirements.txt
Flask==2.1.2
Flask-PyMongo==2.3.0
Werkzeug==2.1.2
pymongo==3.12.1
python-dotenv==0.19.2
pymongo[srv]